+++ freemocap/utilities/remove_empty_directories.py
@@ -1,0 +1,21 @@
+import logging
+from pathlib import Path
+from typing import Union
+
+logger = logging.getLogger(__name__)
+
+
+def remove_empty_directories(root_dir: Union[str, Path]):
+    """
+    Recursively remove empty directories from the root directory
+    :param root_dir: The root directory to start removing empty directories from
+    """
+    # logger.debug(f"Searching for empty directories in: {root_dir}")
+    for path in Path(root_dir).rglob("*"):
+        if path.is_dir() and not any(path.iterdir()):
+            logger.info(f"Removing empty directory: {path}")
+            path.rmdir()
+        elif path.is_dir() and any(path.iterdir()):
+            remove_empty_directories(path)
+        else:
+            continue
